AAAAAAAAAAADDDDDDDDDDDDDDD

+++ core/src/main/java/com/gxa/eloan/common/util/SysConstant.java
@@ -58,20 +58,20 @@ public class SysConstant {
 	// 按月到期还款(每月还利息,到期还本息)
 
 	// ---------------------标的类型--------------------------
-	public final static int BIDREQUEST_TYPE_NORMAL = 0; // 普通信用标
+	public final static Byte BIDREQUEST_TYPE_NORMAL = 0; // 普通信用标
 
 	// ---------------------借款状态---------------------------
-	public final static int BIDREQUEST_STATE_PUBLISH_PENDING = 0; // 待发布
-	public final static int BIDREQUEST_STATE_BIDDING = 1; // 招标中
-	public final static int BIDREQUEST_STATE_UNDO = 2; // 已撤销
-	public final static int BIDREQUEST_STATE_BIDDING_OVERDUE = 3; // 流标
-	public final static int BIDREQUEST_STATE_APPROVE_PENDING_1 = 4; // 满标1审
-	public final static int BIDREQUEST_STATE_APPROVE_PENDING_2 = 5; // 满标2审
-	public final static int BIDREQUEST_STATE_REJECTED = 6; // 满标审核被拒绝
-	public final static int BIDREQUEST_STATE_PAYING_BACK = 7; // 还款中
-	public final static int BIDREQUEST_STATE_COMPLETE_PAY_BACK = 8; // 已还清
-	public final static int BIDREQUEST_STATE_PAY_BACK_OVERDUE = 9; // 逾期
-	public final static int BIDREQUEST_STATE_PUBLISH_REFUSE = 10; // 发标审核拒绝状态
+	public final static Byte BIDREQUEST_STATE_PUBLISH_PENDING = 0; // 待发布
+	public final static Byte BIDREQUEST_STATE_BIDDING = 1; // 招标中
+	public final static Byte BIDREQUEST_STATE_UNDO = 2; // 已撤销
+	public final static Byte BIDREQUEST_STATE_BIDDING_OVERDUE = 3; // 流标
+	public final static Byte BIDREQUEST_STATE_APPROVE_PENDING_1 = 4; // 满标1审
+	public final static Byte BIDREQUEST_STATE_APPROVE_PENDING_2 = 5; // 满标2审
+	public final static Byte BIDREQUEST_STATE_REJECTED = 6; // 满标审核被拒绝
+	public final static Byte BIDREQUEST_STATE_PAYING_BACK = 7; // 还款中
+	public final static Byte BIDREQUEST_STATE_COMPLETE_PAY_BACK = 8; // 已还清
+	public final static Byte BIDREQUEST_STATE_PAY_BACK_OVERDUE = 9; // 逾期
+	public final static Byte BIDREQUEST_STATE_PUBLISH_REFUSE = 10; // 发标审核拒绝状态
 
 	public static final BigDecimal SMALLEST_BID_AMOUNT = new BigDecimal(
 			"50.0000");// 系统规定的最小投标金额
